# finders-keepers-api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/files/upload": {
      "post": {
        "operationId": "FilesController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "entity": {
                    "type": "string",
                    "example": "Product"
                  },
                  "entityId": {
                    "type": "string",
                    "example": "product-id-here"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Files"
        ]
      }
    },
    "/files": {
      "get": {
        "operationId": "FilesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Files"
        ]
      }
    },
    "/files/{id}": {
      "delete": {
        "operationId": "FilesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Files"
        ]
      }
    },
    "/customers": {
      "post": {
        "operationId": "CustomersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Customers"
        ]
      },
      "get": {
        "operationId": "CustomersController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/{id}": {
      "get": {
        "operationId": "CustomersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Customers"
        ]
      },
      "patch": {
        "operationId": "CustomersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/{id}/deactivate": {
      "patch": {
        "operationId": "CustomersController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/{id}/activate": {
      "patch": {
        "operationId": "CustomersController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Customers"
        ]
      }
    },
...
    "/addresses/customer/{customerId}": {
      "get": {
        "operationId": "AddressesController_findByCustomer",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Addresses"
        ]
      }
    },
    "/addresses/{id}": {
      "get": {
        "operationId": "AddressesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Addresses"
        ]
      },
      "patch": {
        "operationId": "AddressesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Addresses"
        ]
      },
      "delete": {
        "operationId": "AddressesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Addresses"
        ]
      }
    },
...
    "/customer-addresses": {
      "post": {
        "operationId": "CustomerAddressesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Customer Addresses"
        ]
      },
      "get": {
        "operationId": "CustomerAddressesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Customer Addresses"
        ]
      }
    },
    "/customer-addresses/{id}": {
      "get": {
        "operationId": "CustomerAddressesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Customer Addresses"
        ]
      },
      "patch": {
        "operationId": "CustomerAddressesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Customer Addresses"
        ]
      },
      "delete": {
        "operationId": "CustomerAddressesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Customer Addresses"
        ]
      }
    },
...
  },
  "info": {
    "title": "Finders Keepers API",
    "description": "Ecommerce backend for Finders Keepers LB",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "customer@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+96170123456"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          }
        }
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "customer@example.com"
          },
          "phone": {
            "type": "string",
            "example": "+96170123456"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          }
        }
      },
...
      "UpdateAddressDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "example": "customer-id-here"
          },
          "label": {
            "type": "string",
            "example": "Home"
          },
          "fullName": {
            "type": "string",
            "example": "John Doe"
          },
          "phone": {
            "type": "string",
            "example": "+96170123456"
          },
          "country": {
            "type": "string",
            "example": "Lebanon"
          },
          "city": {
            "type": "string",
            "example": "Beirut"
          },
          "area": {
            "type": "string",
            "example": "Hamra"
          },
          "street": {
            "type": "string",
            "example": "Main Street"
          },
          "building": {
            "type": "string",
            "example": "Building A"
          },
          "floor": {
            "type": "string",
            "example": "3"
          },
          "apartment": {
            "type": "string",
            "example": "12"
          },
          "notes": {
            "type": "string",
            "example": "Call before delivery"
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "CreateCustomerAddressDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "Home"
          },
          "fullName": {
            "type": "string",
            "example": "John Doe"
          },
          "phone": {
            "type": "string",
            "example": "+96170123456"
          },
          "country": {
            "type": "string",
            "example": "Lebanon"
          },
          "city": {
            "type": "string",
            "example": "Beirut"
          },
          "area": {
            "type": "string",
            "example": "Hamra"
          },
          "street": {
            "type": "string",
            "example": "Main Street"
          },
          "building": {
            "type": "string",
            "example": "Building A"
          },
          "floor": {
            "type": "string",
            "example": "3"
          },
          "apartment": {
            "type": "string",
            "example": "12"
          },
          "notes": {
            "type": "string",
            "example": "Call before delivery"
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "UpdateCustomerAddressDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "Home"
          },
          "fullName": {
            "type": "string",
            "example": "John Doe"
          },
          "phone": {
            "type": "string",
            "example": "+96170123456"
          },
          "country": {
            "type": "string",
            "example": "Lebanon"
          },
          "city": {
            "type": "string",
            "example": "Beirut"
          },
          "area": {
            "type": "string",
            "example": "Hamra"
          },
          "street": {
            "type": "string",
            "example": "Main Street"
          },
          "building": {
            "type": "string",
            "example": "Building A"
          },
          "floor": {
            "type": "string",
            "example": "3"
          },
          "apartment": {
            "type": "string",
            "example": "12"
          },
          "notes": {
            "type": "string",
            "example": "Call before delivery"
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  }
}